Move RoomBadge to atomic package and rename to MatrixBadge
