+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistPreviewTest.kt
@@ -71,6 +71,9 @@ class KonsistPreviewTest {
                 "IconsCompoundPreview",
                 "IconsOtherPreview",
                 "MarkdownTextComposerEditPreview",
+                "MatrixBadgeAtomPositivePreview",
+                "MatrixBadgeAtomNeutralPreview",
+                "MatrixBadgeAtomNegativePreview",
                 "MentionSpanPreview",
                 "MessageComposerViewVoicePreview",
                 "MessagesReactionButtonAddPreview",
@@ -95,9 +98,6 @@ class KonsistPreviewTest {
                 "PollContentViewEndedPreview",
                 "PollContentViewUndisclosedPreview",
                 "ReadReceiptBottomSheetPreview",
-                "RoomBadgePositivePreview",
-                "RoomBadgeNeutralPreview",
-                "RoomBadgeNegativePreview",
                 "RoomMemberListViewBannedPreview",
                 "SasEmojisPreview",
                 "SecureBackupSetupViewChangePreview",
